+++ crates/bonsai-runner/Cargo.toml
@@ -19,6 +19,7 @@ hex = "0.4.3"
 borsh = "1.5.6"
 
 bonsai-sdk = { version = "1.2.4", features = ["non_blocking"] }
+hyle-net = { workspace = true }
 risc0-zkvm = { version = "1.2.4" }
 bytemuck = "1.22.0"
 # Has to be compatible with whatever bonsai-sdk returns, which is not really documented...
