+++ crates/bonsai-runner/src/bin/bonsai-runner.rs
@@ -33,7 +33,9 @@ async fn main() {
     let addr = SocketAddr::from(([127, 0, 0, 1], 3000));
     info!("Server listening on {}", addr);
 
-    let listener = tokio::net::TcpListener::bind("0.0.0.0:3000").await.unwrap();
+    let listener = hyle_net::net::TcpListener::bind("0.0.0.0:3000")
+        .await
+        .unwrap();
     axum::serve(listener, app).await.unwrap();
 }
