+++ crates/client-sdk/Cargo.toml
@@ -9,8 +9,11 @@ repository = { workspace = true }
 rust-version = { workspace = true }
 
 [dependencies]
-sdk = { workspace = true, default-features = false, features = ["full-model"] }
+sdk = { workspace = true, default-features = false, features = [
+    "full-model",
+] }
 bonsai-runner = { workspace = true, optional = true }
+hyle-net = { workspace = true }
 
 anyhow = "1.0.96"
 borsh = "1.5.6"
@@ -27,26 +30,14 @@ reqwest = { version = "0.12", features = ["json"], optional = true }
 
 # Tcp feature
 tokio = { version = "1.42.0", features = ["full", "tracing"], optional = true }
-tokio-util = { version = "0.7.14", optional = true }
-futures = { version = "0.3.31", optional = true }
 
 # Indexer feature, these are re-exported
 axum = { version = "0.8.1", optional = true }
 utoipa = { version = "5.3.1", optional = true }
 utoipa-axum = { version = "0.2.0", optional = true }
-bytes = { version = "1.10.0", optional = true }
-paste = { version = "1.0.15", optional = true }
 
 [features]
 rest = ["dep:reqwest"]
-tcp = [
-  "dep:reqwest",
-  "dep:tokio",
-  "dep:tokio-util",
-  "dep:futures",
-  "dep:bytes",
-  "dep:paste",
-]
 indexer = ["dep:utoipa", "dep:axum", "dep:utoipa-axum", "dep:tokio"]
 risc0 = ["dep:risc0-zkvm", "dep:bonsai-runner"]
 sp1 = ["dep:sp1-sdk", "dep:bincode"]
