+++ crates/client-sdk/src/lib.rs
@@ -3,6 +3,5 @@ pub mod contract_indexer;
 pub mod helpers;
 #[cfg(feature = "rest")]
 pub mod rest_client;
-#[cfg(feature = "tcp")]
-pub mod tcp;
+pub mod tcp_client;
 pub mod transaction_builder;
