+++ crates/hyle-loadtest/Cargo.toml
@@ -14,7 +14,7 @@ hyllar = { workspace = true, features = ["client"] }
 hydentity = { workspace = true, features = ["client"] }
 hyle-contracts = { workspace = true }
 hyle-contract-sdk = { workspace = true, features = ["tracing"] }
-client-sdk = { workspace = true, features= ["rest", "tcp", "risc0"] }
+client-sdk = { workspace = true, features= ["rest", "risc0"] }
 anyhow = "1.0.96"
 borsh = "1.5.6"
 clap = { version = "4.5.31", features = ["derive"] }
