+++ crates/hyle-net/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "hyle-net"
+description = "Hyle network SDK"
+version = { workspace = true }
+edition = { workspace = true }
+# license = { workspace = true }
+homepage = { workspace = true }
+repository = { workspace = true }
+rust-version = { workspace = true }
+
+[dependencies]
+sdk = { workspace = true, default-features = false, features = [
+    "full-model",
+] }
+
+anyhow = "1.0.96"
+borsh = "1.5.5"
+tracing = "0.1"
+
+# Tcp feature
+tokio = { version = "1.42.0", features = ["full", "tracing"], optional = true }
+tokio-util = { version = "0.7.14"}
+futures = { version = "0.3.31" }
+
+bytes = { version = "1.10.0" }
+paste = { version = "1.0.15"}
+
+[features]
+default = ["tokio"]
+tokio = ["dep:tokio"]
