+++ crates/hyle-net/README.md
@@ -1,0 +1,5 @@
+# Hyle net layer
+
+This crates holds some sdk for network communication.
+
+Primarly used to abstract over `tokio::net` and easily substitute it by `turmoil::net`
